refactor: improve file hierarchy and imports
