+++ browser/simulation.js
@@ -10,7 +10,7 @@ import 'aframe-ui-widgets'
 import 'aframe-fence-component'
 import 'aframe-cubemap-component'
 import Sun from './sun'
-import {DriverCam} from './cameras'
+import DriverCam from './cameras'
 import _ from 'lodash'
 import {solution1, solution2, solution3} from './validation'
 import {SpaceshipAmbience} from './soundEffects'
