+++ browser/userCam.jsx
@@ -2,18 +2,8 @@ import 'aframe'
 import {Entity} from 'aframe-react'
 import Tutorial from './tutorial'
 import React from 'react'
-import DayDreamController from './components/dayDreamController'
-
-const Cursor = (props) => (
-  <Entity
-    primitive="a-cursor"
-    animation__click={{
-      property: 'scale',
-      startEvents: 'click',
-      from: '0.1 0.1 0.1',
-      to: '1 1 1',
-      dur: 150}}
-  />)
+import DayDreamController from '../vr/dayDreamController'
+import Cursor from './components/cursor'
 
 const UserCam = (props) => (
     <Entity id='menuCamera' position='0 0 0'>
